MOGL: Fix remaining wrapper function for R2015b+ Matlab bugs.
Hopefully all affected functions are now covered to fix the bug
introduced by Mathworks into Matlab R2015b and later versions.

Hopefully fixes issue #308

+++ Psychtoolbox/PsychOpenGL/MOGL/wrap/gluPickMatrix.m
@@ -15,6 +15,6 @@ if nargin~=5,
     error('invalid number of arguments');
 end
 
-moglcore( 'gluPickMatrix', x, y, delX, delY, viewport );
+moglcore( 'gluPickMatrix', x, y, delX, delY, int32(viewport) );
 
 return
